refactored props function with fp-ts (not working)
