+++ ts/utils/input.ts
@@ -1,6 +1,7 @@
-import { Option } from "fp-ts/lib/Option";
+import { Option, none, some } from "fp-ts/lib/Option";
 import * as t from "io-ts";
 import { PatternString } from "italia-ts-commons/lib/strings";
+import { CreditCard } from "../../definitions/pagopa/CreditCard";
 import { isExpired } from "./dates";
 
 const MIN_PAN_DIGITS = 15;
@@ -40,6 +41,13 @@ export type CreditCardExpirationYear = t.TypeOf<
 export const CreditCardCVC = PatternString("^[0-9]{3,4}$");
 export type CreditCardCVC = t.TypeOf<typeof CreditCardCVC>;
 
+export const INITIAL_CARD_FORM_STATE: CreditCardState = {
+  pan: none,
+  expirationDate: none,
+  securityCode: none,
+  holder: none
+};
+
 export const isValidPan = (pan: Option<string>) =>
   pan.map(pan => CreditCardPan.is(pan)).toUndefined();
 
@@ -59,3 +67,58 @@ export const isValidSecurityCode = (securityCode: Option<string>) =>
   securityCode
     .map(securityCode => CreditCardCVC.is(securityCode))
     .toUndefined();
+
+export type CreditCardState = Readonly<{
+  pan: Option<string>;
+  expirationDate: Option<string>;
+  securityCode: Option<string>;
+  holder: Option<string>;
+}>;
+
+export type CreditCardStateKeys =
+  | "pan"
+  | "expirationDate"
+  | "securityCode"
+  | "holder";
+
+export function mergeCardValues(state: CreditCardState): Option<CreditCard> {
+  const { pan, expirationDate, securityCode, holder } = state;
+  if (
+    pan.isNone() ||
+    expirationDate.isNone() ||
+    securityCode.isNone() ||
+    holder.isNone()
+  ) {
+    return none;
+  }
+
+  const [expirationMonth, expirationYear] = expirationDate.value.split("/");
+
+  if (!CreditCardPan.is(pan.value)) {
+    // invalid pan
+    return none;
+  }
+
+  if (
+    !CreditCardExpirationMonth.is(expirationMonth) ||
+    !CreditCardExpirationYear.is(expirationYear)
+  ) {
+    // invalid date
+    return none;
+  }
+
+  if (!CreditCardCVC.is(securityCode.value)) {
+    // invalid cvc
+    return none;
+  }
+
+  const card: CreditCard = {
+    pan: pan.value,
+    holder: holder.value,
+    expireMonth: expirationMonth,
+    expireYear: expirationYear,
+    securityCode: securityCode.value
+  };
+
+  return some(card);
+}
